+++ quasar-utils/src/test/resources/src/test/java/BasicTest.java
@@ -1,0 +1,4 @@
+public class BasicTest {
+    @org.junit.Test
+    public void doNothing() {}
+}